+++ src/__test__/application/product/remove-discount.test.ts
@@ -1,14 +1,14 @@
 import { faker } from "@faker-js/faker";
 import { describe, expect, test } from "vitest";
-import { ProductPrismaRepository } from "../../../adapters/product/product-prisma-repository";
-import { ShopPrismaRepository } from "../../../adapters/shop/shop-prisma-repository";
-import { UserPrismaRepository } from "../../../adapters/user/user-prisma-repository";
 import { SaveProduct } from "../../../application/product/add-product";
 import { ApplyDiscount } from "../../../application/product/apply-discount";
 import { StoreShop } from "../../../application/shop/store-shop";
 import { StoreAccount } from "../../../application/user/store-account";
 import { VerifyUser } from "../../../application/user/verify-user";
 import { RemoveDiscount } from "../../../application/product/remove-discount";
+import { ShopPrismaRepository } from "../../../infrastructure/db/prisma/shop-prisma-repository";
+import { UserPrismaRepository } from "../../../infrastructure/db/prisma/user-prisma-repository";
+import { ProductPrismaRepository } from "../../../infrastructure/db/prisma/product-prisma-repository";
 
 describe("Remove discount", async () => {
   const input = {
