+++ src/__test__/application/shop/create-shop.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, test } from "vitest";
-import { ShopPrismaRepository } from "../../../adapters/shop/shop-prisma-repository";
 import { StoreShop } from "../../../application/shop/store-shop";
-import { UserPrismaRepository } from "../../../adapters/user/user-prisma-repository";
 import { faker } from "@faker-js/faker";
 import { StoreAccount } from "../../../application/user/store-account";
 import { VerifyUser } from "../../../application/user/verify-user";
+import { ShopPrismaRepository } from "../../../infrastructure/db/prisma/shop-prisma-repository";
+import { UserPrismaRepository } from "../../../infrastructure/db/prisma/user-prisma-repository";
 
 describe("Create shop", async () => {
   const input = {
@@ -60,5 +60,5 @@ describe("Create shop", async () => {
     expect(useCase).toBeDefined();
     expect(useCase).toBeTruthy();
     expect(useCase.statusCode).toBe(201);
-  });
+  }, 10000);
 });
