+++ src/application/user/store-account.ts
@@ -1,9 +1,7 @@
-import { IUserRepository } from "../../adapters/user/user-repository";
 import { Exception } from "../../domain/common/Exception";
 import { StatusCode } from "../../domain/common/status-code";
 import { User } from "../../domain/user/User";
-import { Email } from "../../domain/user/User-email";
-import { Password } from "../../domain/user/User-password";
+import { IUserRepository } from "../../domain/user/user-repository";
 import { HandleReturn } from "../common/handleReturn";
 import { UserExistEmailService } from "./services/user-exist-email";
