+++ src/domain/shop/shop-repository.ts
