+++ src/infrastructure/http/main.ts
@@ -1,8 +1,5 @@
-import { mongoConnect } from "../db/mongodb/connect";
 import { Server } from "./express/Server";
 
 const PORT = 3000;
-
-mongoConnect().catch((err) => console.log(err));
 
 new Server(PORT).start();
